third_party/blink/web_tests/fast/dom/timer-clear-interval-in-handler-and-generate-error-expected.txt

CONSOLE ERROR: Uncaught TypeError: Failed to execute 'appendChild' on 'Node': parameter 1 is not of type 'Node'.
